* Sistem de critico * Fix compilação msvc 2010 * Outras melhorias
git-svn-id: svn://darghos.com.br/darghos/tfs/8.6@3273 5021ead5-42e4-5646-a3ec-22fc1710fdc7

+++ trunk/iologindata.cpp
@@ -332,7 +332,7 @@ void IOLoginData::removePremium(Account account)
 	uint64_t timeNow = time(NULL);
 	if(account.premiumDays > 0 && account.premiumDays < 65535)
 	{
-		uint32_t days = (uint32_t)std::ceil((timeNow - account.lastDay) / 86400);
+		uint32_t days = (uint32_t)std::ceil((float)((timeNow - account.lastDay) / 86400));
 		if(days > 0)
 		{
 			if(account.premiumDays >= days)
